refactor: 변경된 클래스 구조에 따른 테스트 코드 리팩토링
