+++ src/main/java/org/duckstudy/movingcar/Car.java
@@ -5,6 +5,10 @@ public class Car {
     private final String name;
     private long position;
 
+    public long getPosition() {
+        return position;
+    }
+
     public Car(String name) {
         this.name = name;
         this.position = 0;
